Convert in a Maven project
