+++ README.md
@@ -1,4 +1,20 @@
 lotteria
 ========
 
-Simulazione di una lotteria
+Simulation of a lottery.
+
+
+How to build
+------------
+
+Launch the following command in the project directory:
+
+    mvn install
+
+
+Launch
+------
+
+Launch the following command:
+
+    mvn exec:java
