--- src/lotteria/utils/LotteriaUtils.java
@@ -1,16 +1,0 @@
-/**
- * 
- */
-package lotteria.utils;
-
-/**
- * Generic utility for Lotteria.
- * @author <a reef="https://github.com/DavidePastore">DavidePastore</a>
- *
- */
-public class LotteriaUtils {
-	
-	
-	
-
-}
