+++ src/main/java/com/github/davidepastore/lotteria/utils/GeneratoreNomi.java
@@ -1,0 +1,41 @@
+package com.github.davidepastore.lotteria.utils;
+
+import java.io.IOException;
+
+
+/**
+ * It generates names.
+ * @author <a href="https://github.com/DavidePastore">DavidePastore</a>
+ *
+ */
+public class GeneratoreNomi {
+	
+	private static String[] NOMI;
+	private static boolean caricato = false;
+	
+	/**
+	 * Generates a name.
+	 * @return the name
+	 */
+	public static String generaNome(){
+		if(!caricato){
+			caricaNomi();
+			caricato = true;
+		}
+		int randomNumber;
+		randomNumber = GeneratoreNumeri.generaNumero(NOMI.length);
+		return NOMI[randomNumber];
+	}
+	
+	/**
+	 * Load names from the file.
+	 */
+	private static void caricaNomi(){
+		try {
+			NOMI = new FileReader().readLines("Nomi.txt");
+		} catch (IOException e) {
+			System.out.println("Error: "+e);
+		}
+	}
+
+}
