+++ src/main/java/com/github/davidepastore/lotteria/utils/LotteriaUtils.java
@@ -1,0 +1,16 @@
+/**
+ * 
+ */
+package com.github.davidepastore.lotteria.utils;
+
+/**
+ * Generic utility for Lotteria.
+ * @author <a href="https://github.com/DavidePastore">DavidePastore</a>
+ *
+ */
+public class LotteriaUtils {
+	
+	
+	
+
+}
